Updated Evaluation Form View and Metrics Model

+++ TalentAcquisition.BusinessLogic/UpdatedDomain/InterviewEvaluation.cs
@@ -19,11 +19,11 @@ namespace TalentAcquisition.BusinessLogic.UpdatedDomain
         [Display(Name = "Stage ID")]
         public int StageID { get; set; }
         [Display(Name = "Stage 1 Score")]
-        public int Score1 { get; set; }
+        public decimal Score1 { get; set; }
         [Display(Name = "Stage 2 Score")]
-        public int Score2 { get; set; }
+        public decimal Score2 { get; set; }
         [Display(Name = "Stage 3 Score")]
-        public int Score3 { get; set; }
+        public decimal Score3 { get; set; }
         public string ApplicantStrength { get; set; }
         public string ApplicantWeakness { get; set; }
         public string Recommendation { get; set; }
